Fixed some bugs.
Modified fence functions to be added/created a fence with
std::vector<Vertex>.

Added for RTC:
// RTC para medir tiempo real (reloj/calendario)
// Se usa en timer_if.cpp para el RTC.
// NUCLEO STM32WL55JC1 tiene un 32.768 kHz LSE crystal oscillator.
// LSE_VALUE / [(RTC_N_PREDIV_A + 1) * (RTC_N_PREDIV_S + 1)] = tick per
second [Hz]
// 32.768 kHz / [(127 + 1) * (255 + 1)] = 1Hz (un tick x segundo).

+++ CM4/Core/Src/ipcc_if.cpp
@@ -89,7 +89,7 @@ void IPCC_IF_Init(void (*IPCC_IF_ResponseRcv_cb)(uint32_t channelIdx),
   LL_EXTI_EnableRisingTrig_32_63(LL_EXTI_LINE_36);
 
   /* Initialize IPCC callback of all channels*/
-  for (int32_t i = 0; i < IPCC_CHANNEL_NUMBER; i++)
+  for (int32_t i = 0; i < static_cast<int32_t>(IPCC_CHANNEL_NUMBER); i++)
   {
     HAL_IPCC_ActivateNotification(&hipcc, i, IPCC_CHANNEL_DIR_TX, IpccIfIsrTxCb);
     HAL_IPCC_ActivateNotification(&hipcc, i, IPCC_CHANNEL_DIR_RX, IpccIfIsrRxCb);
